@@ -1,5 +1,5 @@
 /**
- * Created by pedro.barreto@bynder.com on 15/Jan/2019.
+ * Created by pedrosousabarreto@gmail.com on 15/Jan/2019.
  */
 "use strict";
 
@@ -8,15 +8,14 @@ import {ServiceParams, ServiceParam, PARAM_TYPES, ServiceFeatureFlag} from "node
 
 let params = new ServiceParams();
 
-
 params.add_param(new ServiceParam("http_port",
-	PARAM_TYPES.NUMBER, 3000,
+	PARAM_TYPES.INT_NUMBER, 3000,
 	"http port for the service to listen on"));
 params.add_param(new ServiceParam("ext_base_url", PARAM_TYPES.STRING,
 	"https://localhost", "external base url, ex: https://localhost:443"));
 
 params.add_param(new ServiceParam("kafka_conn_string", PARAM_TYPES.STRING,
-	"diego.bynder.io:9092", "kafka broker connection string - ex: 127.0.0.1:9092"));
+	"localhost:9092", "kafka broker connection string - ex: 127.0.0.1:9092"));
 
 params.add_param(new ServiceParam("kafka_events_topic", PARAM_TYPES.STRING,
 	"test_events", "topic for test events"));
